Investigate(arm64): pass connectedComponentsWithStatsWithAlgorithm ltype as int (#1984)
Experiment: change the P/Invoke ltype from MatType (a small struct) to int,
matching the native int parameter, for the one confirmed crasher. On the
arm64 register layout this param is the first stack-spilled argument; a small
struct on the arm64 stack is mis-passed on Apple/Windows (not Linux), which
corrupts the trailing returnValue pointer -> garbage deref. Un-skip
ConnectedComponentsTest (which funnels into this extern via ConnectedComponentsEx)
to validate under libgmalloc. Managed-only change; investigation branch only.

+++ src/OpenCvSharp/Cv2/Cv2_imgproc.cs
@@ -2781,7 +2781,7 @@ static partial class Cv2
 
         NativeMethods.HandleException(
             NativeMethods.imgproc_connectedComponentsWithStatsWithAlgorithm(
-                image.ToInputProxy(), labels.ToOutputProxy(), stats.ToOutputProxy(), centroids.ToOutputProxy(), (int)connectivity, ltype, (int)ccltype, out var ret));
+                image.ToInputProxy(), labels.ToOutputProxy(), stats.ToOutputProxy(), centroids.ToOutputProxy(), (int)connectivity, (int)ltype, (int)ccltype, out var ret));
 
         GC.KeepAlive(image);
         GC.KeepAlive(labels);
